third_party/WebKit/LayoutTests/http/tests/activedomobject/marquee.html

<!DOCTYPE HTML>
<html>
<body>
<p>Tests that reparenting marquee elements also reparents ActiveDOMObject.</p>
<iframe src="resources/marquee-frame.html"></iframe>
<script src="resources/activedomobject-test.js"></script>
<script>
    function runTest()
    {
        window.iframe = document.querySelector('iframe').contentDocument;

        log('Before Reparenting');
        shouldBe('window.internals.activeDOMObjectCount(document)', 2);
        shouldBe('window.internals.activeDOMObjectCount(iframe)', 2);

        document.body.appendChild(window.iframe.querySelector('marquee'));

        log('After Reparenting');
        shouldBe('window.internals.activeDOMObjectCount(document)', 2);
        shouldBe('window.internals.activeDOMObjectCount(iframe)', 2);
    }
</script>
</body>
</html>
