third_party/WebKit/LayoutTests/http/tests/activedomobject/marquee-expected.txt

Tests that reparenting marquee elements also reparents ActiveDOMObject.


Before Reparenting
PASS: window.internals.activeDOMObjectCount(document) should be '2' and is.
PASS: window.internals.activeDOMObjectCount(iframe) should be '2' and is.
After Reparenting
PASS: window.internals.activeDOMObjectCount(document) should be '2' and is.
PASS: window.internals.activeDOMObjectCount(iframe) should be '2' and is.
